Infrastructure | Added Test Support Module, Implemented SleepDataConverter & DataFactory

+++ domain/src/main/kotlin/activity/sleep/SleepSession.kt
@@ -5,13 +5,13 @@ import java.time.LocalDateTime
 data class SleepSession(
         val startDate: LocalDateTime,
         val endDate: LocalDateTime,
-        val duration: Double,
+        val duration: Int,
         val isNap: Boolean,
         val sleepQuality: Int,
-        val awakeTime: Double,
-        val remSleep: Double,
-        val lightSleep: Double,
-        val deepSleep: Double,
+        val awakeTime: Int,
+        val remSleep: Int,
+        val lightSleep: Int,
+        val deepSleep: Int,
         val soundsRecorded: Int,
         val mood: Mood
 )
